Split HeadConcludables to be Conclusions of a rule, and ConjunctionConcludables become just Concludables
