+++ logic/Rule.java
@@ -18,25 +18,36 @@
 
 package grakn.core.logic;
 
+import grakn.core.common.exception.GraknException;
 import grakn.core.common.iterator.ResourceIterator;
 import grakn.core.concept.ConceptManager;
 import grakn.core.graph.GraphManager;
 import grakn.core.graph.structure.RuleStructure;
-import grakn.core.logic.concludable.ConjunctionConcludable;
-import grakn.core.logic.concludable.ThenConcludable;
+import grakn.core.logic.concludable.Concludable;
+import grakn.core.logic.tool.ConstraintCopier;
 import grakn.core.pattern.Conjunction;
 import grakn.core.pattern.constraint.Constraint;
+import grakn.core.pattern.constraint.thing.HasConstraint;
+import grakn.core.pattern.constraint.thing.IsaConstraint;
+import grakn.core.pattern.constraint.thing.RelationConstraint;
+import grakn.core.pattern.constraint.thing.ThingConstraint;
+import grakn.core.pattern.constraint.thing.ValueConstraint;
 import grakn.core.pattern.variable.Variable;
 import grakn.core.pattern.variable.VariableRegistry;
 import graql.lang.pattern.Pattern;
 import graql.lang.pattern.variable.ThingVariable;
 
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import static grakn.common.collection.Collections.list;
 import static grakn.common.collection.Collections.set;
+import static grakn.common.util.Objects.className;
+import static grakn.core.common.exception.ErrorMessage.Internal.ILLEGAL_STATE;
+import static grakn.core.common.exception.ErrorMessage.Pattern.INVALID_CASTING;
 import static grakn.core.logic.LogicManager.validateRuleStructureLabels;
 
 
@@ -46,8 +57,8 @@ public class Rule {
     private final RuleStructure structure;
     private final Conjunction when;
     private final Conjunction then;
-    private final Set<ThenConcludable<?, ?>> possibleThenConcludables;
-    private final Set<ConjunctionConcludable<?, ?>> requiredWhenConcludables;
+    private final Set<Conclusion<?>> possibleThenConcludables;
+    private final Set<Concludable<?>> requiredWhenConcludables;
 
     private Rule(LogicManager logicManager, RuleStructure structure) {
         this.logicManager = logicManager;
@@ -59,7 +70,7 @@ public class Rule {
         this.then = thenPattern(structure.then());
         pruneThenTypeHints();
         this.possibleThenConcludables = buildThenConcludables(this.then, this.when.variables());
-        this.requiredWhenConcludables = ConjunctionConcludable.create(this.when);
+        this.requiredWhenConcludables = Concludable.create(this.when);
     }
 
     private Rule(GraphManager graphMgr, ConceptManager conceptMgr, LogicManager logicManager, String label,
@@ -76,7 +87,7 @@ public class Rule {
         pruneThenTypeHints();
 
         this.possibleThenConcludables = buildThenConcludables(this.then, this.when.variables());
-        this.requiredWhenConcludables = ConjunctionConcludable.create(this.when);
+        this.requiredWhenConcludables = Concludable.create(this.when);
         validateCycles();
     }
 
@@ -89,11 +100,11 @@ public class Rule {
         return new Rule(graphMgr, conceptMgr, logicManager, label, when, then);
     }
 
-    public Set<ConjunctionConcludable<?, ?>> whenConcludables() {
+    public Set<Concludable<?>> whenConcludables() {
         return requiredWhenConcludables;
     }
 
-    public Set<ThenConcludable<?, ?>> possibleThenConcludables() {
+    public Set<Conclusion<?>> possibleThenConcludables() {
         return possibleThenConcludables;
     }
 
@@ -181,10 +192,10 @@ public class Rule {
                 });
     }
 
-    private Set<ThenConcludable<?, ?>> buildThenConcludables(Conjunction then, Set<Variable> constraintContext) {
-        HashSet<ThenConcludable<?, ?>> thenConcludables = new HashSet<>();
+    private Set<Conclusion<?>> buildThenConcludables(Conjunction then, Set<Variable> constraintContext) {
+        HashSet<Conclusion<?>> thenConcludables = new HashSet<>();
         then.variables().stream().flatMap(var -> var.constraints().stream()).filter(Constraint::isThing).map(Constraint::asThing)
-                .flatMap(constraint -> ThenConcludable.of(constraint, constraintContext).stream()).forEach(thenConcludables::add);
+                .flatMap(constraint -> Conclusion.of(constraint, constraintContext).stream()).forEach(thenConcludables::add);
         return thenConcludables;
     }
 
@@ -196,4 +207,145 @@ public class Rule {
         return new Conjunction(VariableRegistry.createFromThings(list(thenVariable)).variables(), set());
     }
 
+    public abstract static class Conclusion<CONSTRAINT extends Constraint> {
+
+        private final CONSTRAINT constraint;
+
+        private Conclusion(CONSTRAINT constraint, Set<Variable> constraintContext) {
+            this.constraint = constraint;
+            copyAdditionalConstraints(constraintContext, new HashSet<>(this.constraint.variables()));
+        }
+
+        public static ResourceIterator<? extends Conclusion<?>> of(ThingConstraint constraint, Set<Variable> constraintContext) {
+            Conclusion<?> concludable;
+            if (constraint.isRelation()) concludable = new Relation(constraint.asRelation(), constraintContext);
+            else if (constraint.isHas()) concludable = new Has(constraint.asHas(), constraintContext);
+            else if (constraint.isIsa()) concludable = new Isa(constraint.asIsa(), constraintContext);
+            else if (constraint.isValue()) concludable = new Value(constraint.asValue(), constraintContext);
+            else throw GraknException.of(ILLEGAL_STATE);
+            return null; // TODO Fix
+        }
+
+        public CONSTRAINT constraint() {
+            return constraint;
+        }
+
+        public boolean isRelation() {
+            return false;
+        }
+
+        public boolean isHas() {
+            return false;
+        }
+
+        public boolean isIsa() {
+            return false;
+        }
+
+        public boolean isValue() {
+            return false;
+        }
+
+        public Relation asRelation() {
+            throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Relation.class));
+        }
+
+        public Has asHas() {
+            throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Has.class));
+        }
+
+        public Isa asIsa() {
+            throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Isa.class));
+        }
+
+        public Value asValue() {
+            throw GraknException.of(INVALID_CASTING, className(this.getClass()), className(Value.class));
+        }
+
+
+        private void copyAdditionalConstraints(Set<Variable> fromVars, Set<Variable> toVars) {
+            Map<Variable, Variable> nonAnonfromVarsMap = fromVars.stream()
+                    .filter(variable -> !variable.identifier().reference().isAnonymous())
+                    .collect(Collectors.toMap(e -> e, e -> e)); // Create a map for efficient lookups
+            toVars.stream().filter(variable -> !variable.identifier().reference().isAnonymous())
+                    .forEach(copyTo -> {
+                        if (nonAnonfromVarsMap.containsKey(copyTo)) {
+                            Variable copyFrom = nonAnonfromVarsMap.get(copyTo);
+                            if (copyTo.isThing() && copyFrom.isThing()) {
+                                ConstraintCopier.copyIsaAndValues(copyFrom.asThing(), copyTo.asThing());
+                            } else if (copyTo.isType() && copyFrom.isType()) {
+                                ConstraintCopier.copyLabelAndValueType(copyFrom.asType(), copyTo.asType());
+                            } else throw GraknException.of(ILLEGAL_STATE);
+                        }
+                    });
+        }
+
+        public static class Relation extends Conclusion<RelationConstraint> {
+
+            public Relation(RelationConstraint constraint, Set<Variable> constraintContext) {
+                super(ConstraintCopier.copyConstraint(constraint), constraintContext);
+            }
+
+            @Override
+            public boolean isRelation() {
+                return true;
+            }
+
+            @Override
+            public Relation asRelation() {
+                return this;
+            }
+        }
+
+        public static class Has extends Conclusion<HasConstraint> {
+
+            public Has(HasConstraint constraint, Set<Variable> constraintContext) {
+                super(ConstraintCopier.copyConstraint(constraint), constraintContext);
+            }
+
+            @Override
+            public boolean isHas() {
+                return true;
+            }
+
+            @Override
+            public Has asHas() {
+                return this;
+            }
+        }
+
+        public static class Isa extends Conclusion<IsaConstraint> {
+
+            public Isa(IsaConstraint constraint, Set<Variable> constraintContext) {
+                super(ConstraintCopier.copyConstraint(constraint), constraintContext);
+            }
+
+            @Override
+            public boolean isIsa() {
+                return true;
+            }
+
+            @Override
+            public Isa asIsa() {
+                return this;
+            }
+        }
+
+        public static class Value extends Conclusion<ValueConstraint<?>> {
+
+            public Value(ValueConstraint<?> constraint, Set<Variable> constraintContext) {
+                super(ConstraintCopier.copyConstraint(constraint), constraintContext);
+            }
+
+            @Override
+            public boolean isValue() {
+                return true;
+            }
+
+            @Override
+            public Value asValue() {
+                return this;
+            }
+        }
+    }
 }
