+++ logic/concludable/ConcludableTest.java
@@ -17,6 +17,7 @@
 
 package grakn.core.logic.concludable;
 
+import grakn.core.logic.Rule;
 import grakn.core.pattern.Conjunction;
 import grakn.core.pattern.Disjunction;
 import grakn.core.pattern.constraint.thing.HasConstraint;
@@ -46,26 +47,26 @@ public class ConcludableTest {
         return createFromThings(list(Graql.parseVariable(graqlVariable).asThing())).get(Reference.named(variableName)).asThing();
     }
 
-    private long relationConcludablesCount(Set<ConjunctionConcludable<?, ?>> concludables) {
-        return concludables.stream().filter(ConjunctionConcludable::isRelation).count();
+    private long relationConcludablesCount(Set<Concludable<?>> concludables) {
+        return concludables.stream().filter(Concludable::isRelation).count();
     }
 
-    private long hasConcludablesCount(Set<ConjunctionConcludable<?, ?>> concludables) {
-        return concludables.stream().filter(ConjunctionConcludable::isHas).count();
+    private long hasConcludablesCount(Set<Concludable<?>> concludables) {
+        return concludables.stream().filter(Concludable::isHas).count();
     }
 
-    private long isaConcludablesCount(Set<ConjunctionConcludable<?, ?>> concludables) {
-        return concludables.stream().filter(ConjunctionConcludable::isIsa).count();
+    private long isaConcludablesCount(Set<Concludable<?>> concludables) {
+        return concludables.stream().filter(Concludable::isIsa).count();
     }
 
-    private long valueConcludablesCount(Set<ConjunctionConcludable<?, ?>> concludables) {
-        return concludables.stream().filter(ConjunctionConcludable::isValue).count();
+    private long valueConcludablesCount(Set<Concludable<?>> concludables) {
+        return concludables.stream().filter(Concludable::isValue).count();
     }
 
     @Test
     public void test_conjunction_only_isa_is_built() {
         String conjunction = "{ $a isa b; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(1, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -75,7 +76,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_has_is_built() {
         String conjunction = "{ $a has $b; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(1, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -85,7 +86,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_has_is_built_when_type_and_value_given() {
         String conjunction = "{ $a has age 50; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(1, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -95,7 +96,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_relation_is_built() {
         String conjunction = "{ $a($x, $y); }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(1, relationConcludablesCount(concludables));
@@ -105,7 +106,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_relation_is_built_when_type_is_given() {
         String conjunction = "{ $a($x, $y) isa siblingship; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(1, relationConcludablesCount(concludables));
@@ -115,7 +116,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_has_and_relation_are_built_from_same_owner() {
         String conjunction = "{ $a($x, $y) isa siblingship, has number 2; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(1, hasConcludablesCount(concludables));
         assertEquals(1, relationConcludablesCount(concludables));
@@ -125,7 +126,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_isa_is_built_when_isa_owner_has_value() {
         String conjunction = "{ $a 5; $a isa age; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(1, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -135,7 +136,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_value_is_built() {
         String conjunction = "{ $a 5; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -146,7 +147,7 @@ public class ConcludableTest {
     public void test_conjunction_multiple_has_and_relation_are_built() {
         String conjunction = "{ $a($x, $y, $z) isa transaction, has currency \"GBP\", has value 45; " +
                 "$b(locates: $a); $b has $c; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(3, hasConcludablesCount(concludables));
         assertEquals(2, relationConcludablesCount(concludables));
@@ -156,7 +157,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_multiple_relations_are_built() {
         String conjunction = "{ $a($x); $a($y); }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(2, relationConcludablesCount(concludables));
@@ -166,7 +167,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_only_creates_has_concludable_when_attribute_has_value_constraints() {
         String conjunction = "{ $x has $a; $a isa age; $a > 5; $a <= 10; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(1, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -176,7 +177,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_creates_multiple_value_constraints_for_same_owned_attribute_without_an_isa() {
         String conjunction = "{ $x has $a; $a > 5; $a <= 10; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(1, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -186,7 +187,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_creates_multiple_value_constraints_for_same_attribute() {
         String conjunction = "{ $a > 5; $a <= 10; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(0, isaConcludablesCount(concludables));
         assertEquals(0, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -196,7 +197,7 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_creates_one_has_one_isa_concludable_for_attribute_owning_attribute() {
         String conjunction = "{ $a has $b; $a \"Hi there\" isa content; $b \"English\" isa language; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
         assertEquals(1, isaConcludablesCount(concludables));
         assertEquals(1, hasConcludablesCount(concludables));
         assertEquals(0, relationConcludablesCount(concludables));
@@ -207,8 +208,8 @@ public class ConcludableTest {
     public void test_conjunction_concludables_contain_type_labels() {
         String conjunction = "{ $diag(patient-role-type: $per) isa $diagnosis; $diagnosis type diagnosis-type;" +
                 "$per isa $person; $person type person-type; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
-        RelationConstraint relationConstraint = concludables.stream().filter(ConjunctionConcludable::isRelation).findFirst().get()
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
+        RelationConstraint relationConstraint = concludables.stream().filter(Concludable::isRelation).findFirst().get()
                 .asRelation().constraint();
         IsaConstraint relationIsaConstraint = relationConstraint.owner().isa().get();
         assertEquals("$diagnosis", relationIsaConstraint.type().reference().asName().syntax());
@@ -220,8 +221,8 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_concludables_contain_type_labels_with_anonymous_type_variable() {
         String conjunction = "{ $diag(patient-role-type: $per) isa diagnosis; $per isa person; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
-        RelationConstraint relationConstraint = concludables.stream().filter(ConjunctionConcludable::isRelation).findFirst()
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
+        RelationConstraint relationConstraint = concludables.stream().filter(Concludable::isRelation).findFirst()
                 .get().asRelation().constraint();
         IsaConstraint relationIsaConstraint = relationConstraint.owner().isa().get();
         assertEquals("$_diagnosis", relationIsaConstraint.type().reference().asLabel().syntax());
@@ -232,8 +233,8 @@ public class ConcludableTest {
     @Test
     public void test_conjunction_concludables_contain_type_labels_with_anonymous_type_and_relation_variable() {
         String conjunction = "{ (patient-role-type: $per) isa diagnosis; $per isa person; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
-        RelationConstraint relationConstraint = concludables.stream().filter(ConjunctionConcludable::isRelation).findFirst().get()
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
+        RelationConstraint relationConstraint = concludables.stream().filter(Concludable::isRelation).findFirst().get()
                 .asRelation().constraint();
         IsaConstraint relationIsaConstraint = relationConstraint.owner().isa().get();
         assertEquals("$_diagnosis", relationIsaConstraint.type().reference().asLabel().syntax());
@@ -245,12 +246,12 @@ public class ConcludableTest {
     public void test_conjunction_anonymous_variable_constraints_are_not_conflated() {
         String conjunction = "{ (patient-role-type: $per) isa diagnosis; (patient-role-type: $p2) isa diagnosis; " +
                 "$per isa person; $per has age 50; $per has age 70; }";
-        Set<ConjunctionConcludable<?, ?>> concludables = ConjunctionConcludable.create(parseConjunction(conjunction));
-        concludables.stream().filter(ConjunctionConcludable::isRelation).map(ConjunctionConcludable::asRelation).forEach(relationConcludable -> {
+        Set<Concludable<?>> concludables = Concludable.create(parseConjunction(conjunction));
+        concludables.stream().filter(Concludable::isRelation).map(Concludable::asRelation).forEach(relationConcludable -> {
             assertEquals(0, relationConcludable.constraint().owner().value().size());
             assertTrue(relationConcludable.constraint().owner().isa().isPresent());
         });
-        concludables.stream().filter(ConjunctionConcludable::isHas).map(ConjunctionConcludable::asHas).forEach(hasConcludable -> {
+        concludables.stream().filter(Concludable::isHas).map(Concludable::asHas).forEach(hasConcludable -> {
             assertEquals(1, hasConcludable.constraint().attribute().value().size());
             assertTrue(hasConcludable.constraint().owner().isa().isPresent());
         });
@@ -263,7 +264,7 @@ public class ConcludableTest {
         Conjunction conjunction = parseConjunction("{ $a 5; $a isa age; }");
         ThingVariable variable = parseThingVariable("$a isa age", "a");
         IsaConstraint isaConstraint = variable.isa().get();
-        ThenConcludable.Isa isaConcludable = new ThenConcludable.Isa(isaConstraint, conjunction.variables());
+        Rule.Conclusion.Isa isaConcludable = new Rule.Conclusion.Isa(isaConstraint, conjunction.variables());
         assertEquals(5, isaConcludable.constraint().owner().value().iterator().next().asLong().value().longValue());
     }
 
@@ -272,7 +273,7 @@ public class ConcludableTest {
         Conjunction conjunction = parseConjunction("{ $a > 5; $a isa age; }");
         ThingVariable variable = parseThingVariable("$a isa age", "a");
         IsaConstraint isaConstraint = variable.isa().get();
-        ThenConcludable.Isa isaConcludable = new ThenConcludable.Isa(isaConstraint, conjunction.variables());
+        Rule.Conclusion.Isa isaConcludable = new Rule.Conclusion.Isa(isaConstraint, conjunction.variables());
         assertEquals(5, isaConcludable.constraint().owner().value().iterator().next().asLong().value().longValue());
         assertEquals(GT, isaConcludable.constraint().owner().value().iterator().next().asLong().predicate().asEquality());
     }
@@ -282,7 +283,7 @@ public class ConcludableTest {
         ThingVariable variable = parseThingVariable("$x has $a", "x");
         HasConstraint hasConstraint = variable.has().iterator().next();
         Set<Variable> contextVariables = parseConjunction("{ $x isa $p; $p type person; }").variables();
-        ThenConcludable.Has hasConcludable = new ThenConcludable.Has(hasConstraint, contextVariables);
+        Rule.Conclusion.Has hasConcludable = new Rule.Conclusion.Has(hasConstraint, contextVariables);
         assertEquals("person", hasConcludable.constraint().owner().isa().get().type().label().get().label());
     }
 
@@ -291,7 +292,7 @@ public class ConcludableTest {
         ThingVariable variable = parseThingVariable("$diag(patient: $per) isa $diagnosis", "diag");
         RelationConstraint relationConstraint = variable.relation().iterator().next();
         Set<Variable> contextVariables = parseConjunction("{ $per isa $person; $person type person-type; }").variables();
-        ThenConcludable.Relation relationConcludable = new ThenConcludable.Relation(relationConstraint, contextVariables);
+        Rule.Conclusion.Relation relationConcludable = new Rule.Conclusion.Relation(relationConstraint, contextVariables);
         assertEquals("person-type", relationConcludable.constraint().players().get(0).player().isa().get()
                 .type().label().get().label());
     }
