+++ reasoner/resolution/answer/MappingAggregator.java
@@ -19,7 +19,7 @@ package grakn.core.reasoner.resolution.answer;
 
 import grakn.core.concept.Concept;
 import grakn.core.concept.answer.ConceptMap;
-import grakn.core.logic.concludable.ConjunctionConcludable;
+import grakn.core.logic.concludable.Concludable;
 import graql.lang.pattern.variable.Reference;
 
 import java.util.HashMap;
@@ -44,7 +44,7 @@ public class MappingAggregator extends Aggregator {
         return new MappingAggregator(conceptMap, variableMap);
     }
 
-    public static Map<Reference.Name, Reference.Name> identity(ConjunctionConcludable<?, ?> concludable) {
+    public static Map<Reference.Name, Reference.Name> identity(Concludable<?> concludable) {
         return new HashSet<>(concludable.constraint().variables()).stream()
                 .filter(variable -> variable.reference().isName())
                 .map(variable -> variable.reference().asName())
