+++ reasoner/resolution/resolver/ConjunctionResolver.java
@@ -22,7 +22,7 @@ import grakn.common.collection.Either;
 import grakn.common.collection.Pair;
 import grakn.common.concurrent.actor.Actor;
 import grakn.core.concept.answer.ConceptMap;
-import grakn.core.logic.concludable.ConjunctionConcludable;
+import grakn.core.logic.concludable.Concludable;
 import grakn.core.pattern.Conjunction;
 import grakn.core.reasoner.resolution.MockTransaction;
 import grakn.core.reasoner.resolution.ResolutionRecorder;
@@ -50,11 +50,11 @@ public abstract class ConjunctionResolver<T extends ConjunctionResolver<T>> exte
     private static final Logger LOG = LoggerFactory.getLogger(ConjunctionResolver.class);
 
     final Conjunction conjunction;
-    private final Set<ConjunctionConcludable<?, ?>> conjunctionConcludables;
+    private final Set<Concludable<?>> conjunctionConcludables;
     Actor<ResolutionRecorder> resolutionRecorder;
     private final List<Pair<Actor<ConcludableResolver>, Map<Reference.Name, Reference.Name>>> plannedConcludables;
 
-    public ConjunctionResolver(Actor<T> self, String name, Conjunction conjunction, Set<ConjunctionConcludable<?, ?>> conjunctionConcludables) {
+    public ConjunctionResolver(Actor<T> self, String name, Conjunction conjunction, Set<Concludable<?>> conjunctionConcludables) {
         super(self, name);
 
         this.conjunction = conjunction;
@@ -85,8 +85,8 @@ public abstract class ConjunctionResolver<T extends ConjunctionResolver<T>> exte
         // TODO Tell the concludables to extend themselves by traversing all inconcludable constraints
 
         // Plan the order in which to execute the concludables
-        List<ConjunctionConcludable<?, ?>> planned = list(conjunctionConcludables); // TODO Do some actual planning
-        for (ConjunctionConcludable<?, ?> concludable : planned) {
+        List<Concludable<?>> planned = list(conjunctionConcludables); // TODO Do some actual planning
+        for (Concludable<?> concludable : planned) {
             Pair<Actor<ConcludableResolver>, Map<Reference.Name, Reference.Name>> concludableUnifierPair = registry.registerConcludable(concludable); // TODO TraversalAnswerCount and Rules?
             plannedConcludables.add(concludableUnifierPair);
         }
